Update and rename testing.js to scripts.js
Added multiple functions.  Some of which setup the buttons to be used in order (hiding/showing), which is called upon loading the window.  There is a formula function that takes multiple varying inputs and outputs a single value, with different formulas for different inputs.  I added stopToes and loadToes functions, which stop certain buttons from being seen and show others and changes the strings for the buttons, hiding and showing depending on the existence of a string.  The function forceAmountToes just ensures that textFields have at least 1 input at all times (hopefully).  Lastly, the code is for the most part uncommented, I will get to that alongside bug fixing.